+++ .circleci/scripts/firefox-install.sh
@@ -2,7 +2,7 @@
 
 echo "Installing firefox..."
 sudo rm -r /opt/firefox
-sudo mv firefox /opt/firefox58
+sudo mv firefox /opt/firefox61
 sudo mv /usr/bin/firefox /usr/bin/firefox-old
-sudo ln -s /opt/firefox58/firefox /usr/bin/firefox
+sudo ln -s /opt/firefox61/firefox /usr/bin/firefox
 echo "Firefox installed."