+++ ui/app/components/network-display/network-display.container.js
@@ -1,0 +1,11 @@
+import { connect } from 'react-redux'
+import NetworkDisplay from './network-display.component'
+
+const mapStateToProps = ({ metamask: { network, provider } }) => {
+  return {
+    network,
+    provider,
+  }
+}
+
+export default connect(mapStateToProps)(NetworkDisplay)